convert `scrollContent` to hooks
remove componentDidMount and DidUpdate in favor of useEffects hook.
install and set up react-scroll for smooth scroll in docs. Working great for the main content, not so much for the aside. Persists location on refresh of screen.

WIP: figure out how to have dual scroll so that correct 'is-active' class is positioned at top in aside as well as main article content being positioned accurately on link click.