replacing plea for help resizeObserver
using a resize observer on .Page-content allows us to see when the playground elements find their height. This was the root cause of the scroll issue on refresh. So I am observing the height of the page container until it updates and then disconnecting from it after the first useLayoutEffect is noticed.