refactor of table of contents for readability
this table of contents structure is necessary but i found it hard to read through while updating the styles. I separated out some of the functionality to make it more modular and converted component to function instead of class. Renamed various headings that were getting looped through to show where the separation and looping is actually taking place.