easier to read table of contents
restructuring how the table of contents subsections are formed so that the `renderLinksList` function that creates the jsx for each section just runs 1 time to prevent unnecessary rerenders and use a hook easier.

handleInputChange keys updated so that search inputs are accurate
updated to state hooks

passing isVisible as a prop to styled component to display or hide link according to what is active.