catch up for guides path update
Also back in PR #658 we decided to axe the generic 'guides' (victory-guides) page since it's redundant and you can't get to it unless you manually type it in or go straight from the home page link. What I did not know then and the ci pipeline did not catch is that you cannot simply delete the markdown and call it a day if that route still has children to attend to.
By updating the template of guides to the 404 it's no longer trying to get guide data that doesn't exist and it still allows the children to have the proper url path. the getData is updated to trueDocs so that the proper generic data on the 'docs' template is OK for the children (this is my loose understanding?) - I mostly have just come to the realization that this line is important for things to work and errors to not flash on the screen. Would love to learn more on this or be corrected.