stacking sidenav on top of footer
Moving the header and footer outside of the `main` tag to separate html spaces. This also allows more natural breakage for the fixed header + fixed sidebar with the docs to fill as much space as necessary and then scroll to the footer. By separating the `main` and throwing the nav and the articles in there we can more readily have the required fixed full 100vh aside when on docs and hide it on other pages. I tried a few ways of doing this and to me this was the most legible for future updates and the least convoluted.

While doing this also adjusted the footer spacing and font size to match zeplin docs as well as the padding inconsistency between docs and gallery vs about.